#8343: Update documentation and move sweep test
