+++ .github/workflows/ttnn-run-sweeps.yaml
@@ -80,6 +80,8 @@ on:
           - eltwise.unary.silu.silu
           - eltwise.unary.silu.silu_pytorch2
           - eltwise.unary.glu.glu
+          - eltwise.unary.geglu.geglu
+          - eltwise.unary.swiglu.swiglu
           - eltwise.unary.sigmoid.sigmoid
           - eltwise.unary.sigmoid.sigmoid_pytorch2
           - eltwise.unary.sigmoid_accurate.sigmoid_accurate
